requirements.txt
django==2.2.2
djangorestframework==3.9.4
pdfkit==0.6.1
requests==2.22.0
gunicorn==19.9.0
